pyproject.toml
[project]
name = "symphony"
version = "0.1.0"
description = "Autonomous coding agent orchestrator for Linear and GitHub issues using Codex app-server"
readme = "README.md"
authors = [
    { name = "Manoj Bajaj", email = "manojbajaj95@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-liquid>=2.1.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]

[project.scripts]
symphony = "symphony:main"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
    "ty>=0.0.1a12",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
